@@ -1,16 +1,28 @@
 module btc-alert
 
-go 1.16
+go 1.20
 
 require (
 	github.com/PuerkitoBio/goquery v1.6.1
-	github.com/bwmarrin/discordgo v0.23.2
-	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
+	github.com/blend/go-sdk v1.20220411.3
+	github.com/bwmarrin/discordgo v0.27.1
 	github.com/gorilla/mux v1.8.0
+	github.com/jedib0t/go-pretty/v6 v6.5.6
 	github.com/olekukonko/tablewriter v0.0.5
+	github.com/samber/lo v1.39.0
 	github.com/sirupsen/logrus v1.8.1
-	github.com/t-tomalak/logrus-easy-formatter v0.0.0-20190827215021-c074f06c5816
-	github.com/tsny/btc-alert v0.0.0-20210514033756-2697518b44e1
 	github.com/wcharczuk/go-chart v2.0.1+incompatible
+)
+
+require (
+	github.com/andybalholm/cascadia v1.1.0 // indirect
+	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
+	github.com/gorilla/websocket v1.4.2 // indirect
+	github.com/mattn/go-runewidth v0.0.15 // indirect
+	github.com/rivo/uniseg v0.2.0 // indirect
+	golang.org/x/crypto v0.0.0-20220112180741-5e0467b6c7ce // indirect
+	golang.org/x/exp v0.0.0-20220303212507-bbda1eaf7a17 // indirect
 	golang.org/x/image v0.0.0-20210504121937-7319ad40d33e // indirect
+	golang.org/x/net v0.0.0-20211216030914-fe4d6282115f // indirect
+	golang.org/x/sys v0.16.0 // indirect
 )
