@@ -3,46 +3,48 @@ package eps
 const capLimit = 10000
 
 type CandleQueue struct {
-	Oldest Candlestick
-	Newest Candlestick
-	inner  []Candlestick
+	Newest *Candlestick
+	Oldest *Candlestick
+	Array  []*Candlestick
 }
 
-func NewQueue() *CandleQueue {
-	return &CandleQueue{}
-}
-
-func (q *CandleQueue) Subscribe(p *Publisher) *CandleQueue {
-	p.RegisterSubscriber(q.handlePriceUpdate)
+func NewQueue(p *Publisher) *CandleQueue {
+	q := &CandleQueue{}
+	q.Subscribe(p)
 	return q
 }
 
-func (q *CandleQueue) handlePriceUpdate(p *Publisher, c Candlestick) {
+// Subscribe makes the queue listen for the publisher's events/updates
+func (q *CandleQueue) Subscribe(p *Publisher) {
+	p.RegisterPriceUpdateListener(q.handlePriceUpdate)
+}
+
+func (q *CandleQueue) handlePriceUpdate(p *Publisher, c *Candlestick, completed bool) {
 	// Ignore incomplete candles
-	if !c.Complete {
+	if !completed {
 		return
 	}
-	q.Add(c)
+	q.Enqueue(c)
 }
 
 func (q *CandleQueue) GetAllPrices() []float64 {
 	var arr []float64
-	for _, v := range q.inner {
+	for _, v := range q.Array {
 		arr = append(arr, v.Close)
 	}
 	return arr
 }
 
-func (q *CandleQueue) Add(c Candlestick) {
+func (q *CandleQueue) Enqueue(c *Candlestick) {
 	// todo: make this better than just truncating
 	// this is actually backwards right now which looks fine in graphs
-	if len(q.inner) > capLimit {
-		q.inner = q.inner[:capLimit-1]
+	if len(q.Array) > capLimit {
+		q.Array = q.Array[:capLimit-1]
 	}
-	q.inner = append(q.inner, c)
+	q.Array = append(q.Array, c)
 	q.Newest = c
 }
 
-func (q *CandleQueue) GetQueue() []Candlestick {
-	return q.inner
+func (q *CandleQueue) GetQueue() []*Candlestick {
+	return q.Array
 }
