@@ -1,42 +1,35 @@
 package main
 
 import (
+	"fmt"
 	"net/http"
 	"os"
+	"strings"
 
-	"btc-alert/binance"
 	"btc-alert/coinbase"
 	"btc-alert/eps"
-	"btc-alert/yahoo"
 
+	"github.com/samber/lo"
 	log "github.com/sirupsen/logrus"
 )
 
 var lookupService = eps.NewSecurityLookup()
+var publishers = []*eps.Publisher{}
+var queues = map[string]*eps.CandleQueue{}
 
 func main() {
 
+	if len(os.Args) > 1 {
+		c := coinbase.Get24Hour(os.Args[1])
+		fmt.Printf("%+v\n", c)
+		return
+	}
+	readConfig()
+
 	// Crypto
-	for _, sec := range coinbase.CryptoMap {
-		pub := eps.NewPublisher(coinbase.GetPrice, sec.Ticker, "Coinbase", false, 30)
-		go trackSecurity(pub, sec)
-	}
-
-	// Stocks
-	for _, ticker := range conf.YahooTickers {
-		go func(ticker string) {
-			pub := eps.NewPublisher(yahoo.GetPrice, ticker, "Yahoo", false, 30)
-			pub.UseMarketHours = true
-			name := yahoo.GetDetails(ticker).ShortName
-			sec := eps.NewStock(name, ticker, "Yahoo")
-			go trackSecurity(pub, sec)
-		}(ticker)
-	}
-
-	// DOGE
-	pub := eps.NewPublisher(binance.GetPrice, "DOGEUSDT", "Binance", false, 30)
-	sec := eps.NewCrypto("DOGE", "DOGEUSDT", "Yahoo", "DOGE", "DOGECOIN")
-	go trackSecurity(pub, sec)
+	btcPublisher := eps.NewPublisher(coinbase.GetPrice, coinbase.BTC, "Coinbase", false, 60, 20)
+	publishers = append(publishers, btcPublisher)
+	go track(btcPublisher)
 
 	// API Init
 	port := os.Getenv("PORT")
@@ -45,34 +38,27 @@ func main() {
 		log.Infof("Defaulting to port %s", port)
 	}
 
-	trackGainers()
 	log.Infof("Listening on port %s", port)
 	log.Fatal(http.ListenAndServe(":"+port, initRoutes()))
 }
 
-func trackSecurity(pub *eps.Publisher, sec *eps.Security) *eps.InfoBall {
-	_ = newListener(pub, conf.Intervals, conf.Thresholds)
-	queue := eps.NewQueue().Subscribe(pub)
-	info := lookupService.Register(sec, pub, queue)
+func track(pub *eps.Publisher) {
+	// _ = newListener(pub, conf.Intervals, conf.Thresholds)
+	queue := eps.NewQueue(pub)
+	queues[pub.Ticker] = queue
 	pub.SetActive(true)
-	return info
 }
 
-// Finds the top gainers today from Yahoo and follows them
-func trackGainers() {
-	if !eps.IsMarketHours() {
-		log.Warn("ignoring call to trackGainers as it is not market hours")
-		return
+func findPublisher(s string) (*eps.Publisher, bool) {
+	return lo.Find(publishers, func(e *eps.Publisher) bool {
+		return strings.Contains(strings.ToLower(e.Ticker), s)
+	})
+}
+
+func findQueue(ticker string) (*eps.CandleQueue, bool) {
+	key, ok := lo.Find(lo.Keys(queues), func(s string) bool { return strings.Contains(s, ticker) })
+	if !ok {
+		return nil, false
 	}
-	gainers := yahoo.GetGainers()
-	for _, v := range gainers {
-		if info := lookupService.FindSecurityByNameOrTicker(v); info == nil {
-			if deets := yahoo.GetDetails(v); deets != nil && deets.ShortName != "" {
-				sec := eps.NewStock(deets.ShortName, v, "Yahoo")
-				pub := eps.NewPublisher(yahoo.GetPrice, v, "Yahoo", true, 30)
-				pub.UseMarketHours = true
-				go trackSecurity(pub, sec)
-			}
-		}
-	}
+	return queues[key], true
 }
